move FastXReaderSupport trait, and implement inside fasta/fastq to avoid pub(crate)
