+++ src/fastq.rs
@@ -3,7 +3,7 @@ use std::io;
 #[cfg(feature = "niffler")]
 use std::path::Path;
 
-use crate::{Error, Record, DEFAULT_MAX_RECORDS};
+use crate::{fastx::FastXReaderSupport, Error, Record, DEFAULT_MAX_RECORDS};
 
 pub struct Reader<R: io::Read> {
     /// Handle to the underlying reader (byte stream)
@@ -15,7 +15,7 @@ pub struct Reader<R: io::Read> {
     /// Sets the maximum capcity of records in batches for parallel processing
     ///
     /// If not set, the default `RecordSet` capacity is used.
-    pub(crate) batch_size: Option<usize>,
+    batch_size: Option<usize>,
 }
 
 #[cfg(feature = "niffler")]
@@ -221,13 +221,13 @@ impl<R: io::Read> Reader<R> {
 #[derive(Debug)]
 pub struct RecordSet {
     /// Main buffer for records
-    pub(crate) buffer: Vec<u8>,
+    buffer: Vec<u8>,
     /// Store newlines in buffer
     newlines: Vec<usize>,
     /// Track the last byte position we've searched for newlines
     last_searched_pos: usize,
     /// Position tracking for complete records
-    pub(crate) positions: Vec<Positions>,
+    positions: Vec<Positions>,
     /// Maximum number of records to store
     capacity: usize,
     /// Average number of bytes per record
@@ -392,7 +392,7 @@ impl RecordSet {
 }
 
 #[derive(Debug, Default, Clone, Copy)]
-pub(crate) struct Positions {
+struct Positions {
     start: usize,
     seq_start: usize,
     sep_start: usize,
@@ -406,7 +406,7 @@ pub struct RefRecord<'a> {
     positions: Positions,
 }
 impl<'a> RefRecord<'a> {
-    pub(crate) fn new(buffer: &'a [u8], positions: Positions) -> Result<Self, Error> {
+    fn new(buffer: &'a [u8], positions: Positions) -> Result<Self, Error> {
         let ref_record = Self { buffer, positions };
         ref_record.validate_record()?;
         Ok(ref_record)
@@ -497,6 +497,37 @@ impl Record for RefRecord<'_> {
         Some(self.access_buffer(self.positions.qual_start, self.positions.end))
     }
 }
+
+impl<R> FastXReaderSupport for crate::fastq::Reader<R>
+where
+    R: io::Read + Send,
+{
+    type RecordSet = crate::fastq::RecordSet;
+    type Error = crate::Error;
+    type RefRecord<'a> = crate::fastq::RefRecord<'a>;
+
+    fn new_record_set(&self) -> Self::RecordSet {
+        if let Some(batch_size) = self.batch_size {
+            Self::RecordSet::new(batch_size)
+        } else {
+            Self::RecordSet::default()
+        }
+    }
+
+    fn fill(&mut self, record: &mut Self::RecordSet) -> std::result::Result<bool, Self::Error> {
+        record.fill(self)
+    }
+
+    fn iter(
+        record_set: &Self::RecordSet,
+    ) -> impl Iterator<Item = std::result::Result<Self::RefRecord<'_>, crate::Error>> {
+        record_set
+            .positions
+            .iter()
+            .map(move |&pos| Self::RefRecord::new(&record_set.buffer, pos))
+    }
+}
+
 
 #[cfg(test)]
 mod tests {
